% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhr_data.R
\name{dhr_data}
\alias{dhr_data}
\title{Find the file path to the latest dataset}
\usage{
dhr_data(data_source)
}
\arguments{
\item{data_source}{A single string (character) that relates to one of the specification names. This must match the specification name exactly.}
}
\value{
A character string containing the full file path of the desired file, including the server name. This requires that the user has access to the shared drives on the network. This function utilizes the check_folder_access() function to ensure that the analyst has access to the data.
}
\description{
This allows the user to obtain the full file path to the latest data. This function only works on finding the most current data and does not work on finding older datasets.
}
\examples{
jpact_file <- dhr_data("JPACT")
jpact_file  # print the file path
}
