% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_specs.R
\name{import_specs}
\alias{import_specs}
\title{Text file specifications for weekly extracts}
\usage{
import_specs(spec = "all", keep = NULL, drop = NULL)
}
\arguments{
\item{spec}{A single string (character) that relates to one of the specification names. This must match the specification name exactly.}

\item{keep}{A character vector (e.g., c('EmployeeID', 'Birthdate', 'TitleCode')) that allows the user to keep only specific fields (columns) when importing a text file. This is useful when there are many fields that are unneeded. Defaults to keep all.}

\item{drop}{A character vector that allows the user to drop certain fields when importing a text file. This is useful when there are only a few fields that are unwanted. Defaults to NULL, which effectively keeps all fields when importing.}
}
\value{
By default, import_specs() with no arguments specified returns a data.frame object containing all the names of the text files and their corresponding fields, start points, end points, and field widths. The resulting data.frame has columns SpecID, SpecName, FieldName, Start, Width, and End. By using the keep or drop arguments, you are effectively filtering this data.frame for the fields you want to keep or drop.
}
\description{
Returns the full table of specifications for fixed-width text files that are extracted weekly. You can explore the table for a full list of data fields that are available. The resulting table is designed to be used with the readr package to read text files.
}
\examples{
## Not run:
require(dplyr)
require(stringr)
demo_specs <- import_specs("EmplDemogr")
demo_file <- dhr_data("EmplDemogr")

require(readr)

demo_data <- readr::read_fwf( demo_file,
                      col_positions = readr::fwf_positions(
                                      start = demo_specs$Start,
                                      end = demo_specs$End ,
                                      col_names = demo_specs$FieldName))
}
