% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_filter.R
\name{my_filter}
\alias{my_filter}
\title{The core filter function}
\usage{
my_filter(df, filt_list)
}
\arguments{
\item{df}{The data.frame to which the criteria list is applied.}

\item{filt_list}{A list of criteria.}
}
\value{
This function returns a data.frame that has been filtered.
}
\description{
This function is used by the other functions in this package to apply the filter criteria provided by the user/analyst.
}
