% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_search.R
\name{record_search}
\alias{record_search}
\title{Filter employee records}
\usage{
record_search(df, filter_list)
}
\arguments{
\item{df}{The data.frame to which the criteria list is applied.}

\item{filter_list}{A list of criteria.}
}
\value{
A data.frame that has been filtered to show the most recent employee records. Use this function to find current employees based on any number of criteria. The only criteria that must be provided is the EXPIRATION_DT being equal to 12319999.
}
\description{
This function allows users/analysts to filter a data.frame using an arbitrary list of criteria. This function and other filtering functions in this package are based on the JPACT extract, which contains a few of the required column names (e.g., APPOINTMENT_ID, EMPLOYEE_ID)
}
\examples{
## Not run:
require(lubridate)
# Find the queried records and then get a count
jpact \%>\%
   record_search(
       list(
           EXPIRATION_DT = c(mdy("12319999")) ,
           HOME_DEPT_CD = grep("GJ|NL|SC|NB", unique(jpact$HOME_DEPT_CD), value = TRUE, invert = TRUE),
           EMPLMT_STA_CD = c("A") ,
           SUB_TITLE_CD = c("A", "D", "L", "N")
       )
   ) \%>\%
   nrow()
}
