% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spread_n.R
\name{spread_n}
\alias{spread_n}
\title{Spread any number of variables as new columns in a data.frame}
\usage{
spread_n(df, key, value, fill)
}
\arguments{
\item{df}{data.frame}

\item{key}{The new spread-out column names will be taken from this specified column}

\item{value}{What values (i.e., data) should be filled in these new columns?}

\item{fill}{How should missing values be filled? Usually '0'.}
}
\value{
A wider data.frame than the original.
}
\description{
This expands tidyr::spread() to take on any number of variables and spread them out all at once.
}
\examples{
require(tidyr)
require(dplyr)

set.seed(100)

df <- tibble( Dept = c(rep("DeptA", 2), rep("DeptB", 2)),
              Gender = rep(c("Male", "Female"), 2),
              Count = sample(1:10, 4, replace=T),
              Percent = runif(4, 20, 80) )

df \%>\% spread_n( key = Gender, value = c(Count, Percent), fill = 0 )
}
